fix: discord notification data
