+++ package.json
@@ -11,6 +11,8 @@
     "postinstall": "nuxt prepare",
     "add:automation": "pnpm exec tsx scripts/parse-automation-issue.ts",
     "scan:users:hourly": "pnpm exec tsx --env-file=.env scripts/scan-users.ts --output=hourly-scan-results.txt --max-scans=25 --window-output=hourly-window-scan-results.txt --window-max-scans=30 --daily-output=daily-scan-results.json",
+    "trigger:discord": "pnpm exec tsx --env-file=.env scripts/notify-discord.ts",
+    "trigger:discord:dry-run": "pnpm exec tsx --env-file=.env scripts/notify-discord.ts --dry-run",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "format": "prettier --write .",
